Add files via upload
 Introduction and Objective

The automotive industry is extremely competitive. With increasing fuel prices and picky consumers, automobile makers are constantly optimizing their processes to increase fuel efficiency. 
The objective of this project is to study the relationship between Displacement, Horsepower, Cylinders, Acceleration, Weight and Miles Per Gallon(Mpg) and establish a model which would successfully predict fuel efficiency i.e Mpg of a car with a good accuracy.